requirements.txt
django>=1.2.3
-e hg+http://bitbucket.org/andrewgodwin/south/@0.7.1#egg=django_south
-e svn+http://django-compress.googlecode.com/svn/trunk/#egg=django_compress
-e git+http://github.com/django-extensions/django-extensions.git#egg=django-extensions
-e git+http://github.com/robhudson/django-debug-toolbar.git#egg=django-debug-toolbar
-e svn+http://django-syncr.googlecode.com/svn/trunk/#egg=syncr
python-memcached
MySQL-python
gunicorn
django-sentry